move the location helpers to their own file
